### 2024-01-26 4a466e3
updated docs and setup.py
