+++ docs/build/html/_sources/installation.rst
@@ -1,0 +1,7 @@
+============
+Installation
+============
+
+You can access the latest update of the modulo python package on PyPI using the command line::
+
+    $ pip install modulo-vki
