+++ docs/build/_sources/installation.rst.txt
@@ -1,0 +1,7 @@
+============
+Installation
+============
+
+You can access the latest update of the modulo python package on PyPI. At the command line::
+
+    $ pip install modulo-vki
