+++ docs/build/_sources/intro.rst.txt
@@ -1,0 +1,77 @@
+=====================
+MODULO: Introduction
+=====================
+
+MODULO (MODal mULtiscale pOd) is a software developed at the von Karman Institute toperform data-driven modal decompositions.
+Initially focused on the Multiscale Proper Orthogonal Decomposition (mPOD), it has recently been extended to perform also other decompositions.
+
+.. image:: modulo_logo.png
+  :target: https://imgur.com/ylm0cyv
+  :width: 700
+  :align: center
+  :alt: Alternative text
+
+++++++++++++++++++++++++++++++++++++++++++++++++++++
+What are data-driven decompositions? Some literature
+++++++++++++++++++++++++++++++++++++++++++++++++++++
+
+These seeks to partition a dataset as a linear combination of elementary contributions called modes.
+For a brief overview of the context, we refer to: 
+
+- Ninni, D., Mendez, M. A. (2020), "MODULO: A Software for Multiscale Proper Orthogonal Decomposition of data", Software X, Vol 12, 100622, https://doi.org/10.1016/j.softx.2020.100622.
+
+- Poletti, R., Schena, L., Ninni, D., Mendez, M.A (2024) "MODULO: a python toolbox for data-driven modal decomposition", Submitted to Journal of Open Source Software. Preprint available at https://www.researchgate.net/publication/376885484_MODULO_a_python_toolbox_for_data-driven_modal_decomposition
+
+The first article also presents the first version of MODULO (available in the OLD_Matlab_Implementation branch) and its GUI developed by D. Ninni. The second introduces MODULO v2 in this branch and alternative open source projects.
+While many projects allows for computing common decompositions such as POD, DMD and the SPODs, MODULO is currently the only opensource project allowing to compute the mPOD.
+
+For a more comprehensive overview on the theory of data-driven decompositions, we refer to the chapter:
+
+- Mendez, M. A. (2023) "Generalized and Multiscale Modal Analysis". In : Mendez M.A., Ianiro, A., Noack, B.R., Brunton, S. L. (Eds), "Data-Driven Fluid Mechanics: Combining First Principles and Machine Learning". Cambridge University Press, 2023:153-181. https://doi.org/10.1017/9781108896214.013. The pre-print is available at https://arxiv.org/abs/2208.12630. 
+
+and the article that first presented the complete theory of the mPOD :
+
+- Mendez, M. A., Balabane, M., Buchlin, J.-M. (2019) "Multi-Scale Proper Orthogonal Decomposition of Complex Fluid Flows" Journal of Fluid Mechanics 870:988-1036, https://doi.org/10.1017/9781108896214.013. The pre-print is available at https://arxiv.org/abs/2208.12630. 
+
+The first version of the mPOD was based on wavelet decomposition and was presented here :
+
+- Mendez, M.A.,  Scelzo, M.T., Buchlin, J.-M. , Multiscale Modal Analysis of an Oscillating Impinging Gas Jet, Experimental Thermal and Fluid Science, Vol 91, February 2018, pp. 256-276 (https://doi.org/10.1016/j.expthermflusci.2017.10.032).
+
+Ongoing works on nonlinear methods are discussed here:
+
+- Mendez, M. A. (2023) "Linear and Nonlinear Dimensionality Reduction from Fluid Mechanics to Machine Learning", Meas. Sci. Technol. 34(042001), https://doi.org/10.1088/1361-6501/acaffe. The pre-print is available at https://arxiv.org/abs/2208.07746.   
+
+Some examples of applications of mPOD for processing experimental and numerical are listed below:
+
[…]
