+++ modulo_vki.egg-info/SOURCES.txt
@@ -3,6 +3,21 @@ README.md
 setup.py
 modulo/__init__.py
 modulo/modulo.py
+modulo/core/__init__.py
+modulo/core/_dft.py
+modulo/core/_dmd_s.py
+modulo/core/_k_matrix.py
+modulo/core/_mpod_space.py
+modulo/core/_mpod_time.py
+modulo/core/_pod_space.py
+modulo/core/_pod_time.py
+modulo/core/_spod_s.py
+modulo/core/_spod_t.py
+modulo/utils/__init__.py
+modulo/utils/_plots.py
+modulo/utils/_utils.py
+modulo/utils/others.py
+modulo/utils/read_db.py
 modulo_vki.egg-info/PKG-INFO
 modulo_vki.egg-info/SOURCES.txt
 modulo_vki.egg-info/dependency_links.txt
