### 2024-04-03 #6208 @mendezVKI
Hello again Philip!

The toolbox published in https://www.softxjournal.com/article/S2352-7110(20)30335-6/fulltext was the first version of the toolbox in this submission. It was written in Matlab and equipped with a GUI and an executable.  That version is no longer maintained and has been kept in one of the branches of the MODULO repository.

The version we have submitted to JOSS has been entirely rewritten in Python, updated with modern packages from scipy/numpy and considerably extended to have many more decompositions (the SPODs, the DMD, the Kernel POD) and features (non-uniform grids, more sophisticated memory saving, etc.).

The main novelties are listed on the doc page: https://modulo.readthedocs.io/en/latest/what_is_new.html .

The reason why I kept the original repo and did not move to a new one is that I have promoted it in various published materials (most notably in this chapter: https://www.cambridge.org/core/books/abs/datadriven-fluid-mechanics/generalized-and-multiscale-modal-analysis/F6C342E80A83D68756CEC1492848C6E1 ) and I would like to preserve that.
