### 2024-04-08 #6208 @philipcardiff
Hi @modelflows, Would you be interested in reviewing the submission “MODULO: a python toolbox for data-driven modal decomposition ”, which we received at JOSS? The link for the submission is https://github.com/openjournals/joss-reviews/issues/6208, and the repository is at https://github.com/mendezVKI/MODULO.

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. You can have a look at our review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html) and review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected.
