+++ docs/source/installation.rst
@@ -2,6 +2,6 @@
 Installation
 ============
 
-You can access the latest update of the modulo python package on PyPI. At the command line::
+You can access the latest update of the modulo python package on PyPI using the command line::
 
     $ pip install modulo-vki
