### 2024-04-17 1d3c59e
2.0.4: updated docs w/ new name, fixed imports in tutorial, pub. to pypi
