+++ docs/build/_sources/usage.rst.txt
@@ -8,8 +8,11 @@ to perform the POD.
 
 .. code-block:: python
 
-    from modulo.modulo import MODULO    
-    m = MODULO(data=D); Phi_POD, Psi_POD, Sigma_POD = m.compute_POD_K()
+    from modulo_vki import ModuloVKI  
+    m = ModuloVKI(data=D)
+	Phi_POD, Psi_POD, Sigma_POD = m.compute_POD_K()
+
+
 
 
 
