+++ docs/source/usage.rst
@@ -8,9 +8,11 @@ to perform the POD.
 
 .. code-block:: python
 
-    from modulo.modulo import MODULO    
-    m = MODULO(data=D)
+    from modulo_vki import ModuloVKI  
+    m = ModuloVKI(data=D)
 	Phi_POD, Psi_POD, Sigma_POD = m.compute_POD_K()
+
+
 
 
 
