+++ modulo_vki.egg-info/requires.txt
@@ -7,5 +7,3 @@ ipython
 ipython-genutils
 ipywidgets
 matplotlib
-pyvista
-imageio

[…]
