### 2024-05-06 e0ae3aa
Path fixes in tutorials 4 and 5
