+++ examples/python_files/ex_05_nonUniform_POD.py
@@ -176,13 +176,13 @@ pl.show()
 from modulo_vki import ModuloVKI
 
 if N_PARTITIONS==1:
-    m = ModuloVKI(data=D, n_Modes=5, dtype='float32' ,weights =weights_for_D)
+    m = ModuloVKI(data=D, n_Modes=5, dtype='float32' ,weights =weights_for_D,FOLDER_OUT=FOLDER+os.sep+'MODULO_tmp')
     Phi_POD, Psi_POD, Sigma_POD = m.compute_POD_K()
 else:
     # Prepare 10 partitions, see ex_04 for more details
-    _ = ReadData._data_processing(D=D,N_PARTITIONS=N_PARTITIONS,FOLDER_OUT='./MODULO_tmp')
+    _ = ReadData._data_processing(D=D,N_PARTITIONS=N_PARTITIONS,FOLDER_OUT=FOLDER+os.sep+'MODULO_tmp')
     # Make sure to give the dimensions of D as input to MODULO when D is no more saved in the RAM
-    m = ModuloVKI(data=None, n_Modes=5, dtype='float32', weights =weights_for_D,N_PARTITIONS=N_PARTITIONS,N_S=N_P*N,N_T=N_T)
+    m = ModuloVKI(data=None, n_Modes=5, dtype='float32', weights =weights_for_D,N_PARTITIONS=N_PARTITIONS,N_S=N_P*N,N_T=N_T,FOLDER_OUT=FOLDER+os.sep+'MODULO_tmp')
     Phi_POD, Psi_POD, Sigma_POD = m.compute_POD_K()
 
 #%%
