### 2024-05-15 #6753 @editorialbot
[REVIEW]: MODULO: a python toolbox for data-driven modal decomposition
**Submitting author:** @mendezVKI (<a href="http://orcid.org/0000-0002-1115-2187">Miguel Alfonso Mendez</a>)
**Repository:** https://github.com/mendezVKI/MODULO
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.7
