### 2024-06-25 #6753 @philipcardiff
Hi @mengaldo, If anything about the review process is unclear, just let me know. You can generate your checklist by creating a comment with the command `@editorialbot generate my checklist`
