### 2024-06-27 mendezVKI/MODULO#27 @AndreWeiner
JOSS review
Hi @mendezVKI

This issue is related to the active review process going on [here](https://github.com/openjournals/joss-reviews/issues/6753). In general, the repository is in good shape. A few open points should to be addressed before I can check the remaining boxes.

Regarding the article:
- [x] overall, the writing is good; there are a few typos here and there; I'd recommend running the manuscript through a tool like Grammarly or similar; here are a few lines to check:
  - line 18: computing, line 20: data-driven, line 24: large dimensional (sounds odd to me), line 30: the modes the ones, line 40: bringing POD and DFT, line 50: to perform
- [x] I suggest including an mPOD example in the article since this method is the main argument for this package; ideally, the example should highlight the main advantage/difference of mPOD compared to related methods, e.g., standard POD
- [x] I think the *statement of need* would benefit from a more thorough comparison between the current and related packages (PyDMD, PySPOD, [ModRed](https://github.com/belson17/modred))

Regarding the tutorials:
- [x] tutorial 1 notebook: the saved eigenvalue plots are empty due to calling plt.show() before plt.savefig()
- [x] tutorial 2 notebook: there is a left-over comment *can you check that?*
- [x] tutorial 3 notebook: `D_MEAN` is used before defined when executing the cells in order
- [x] tutorial 4 notebook: `MODULOVKI` causes a crash and seems to be an outdated module name (`ModuloVKI` works); also, the plots containing the time coefficients are empty
- [x] the plotting issues also appear in the related Python scripts; personally, I find the matplotlib pop-up windows a little bit annoying, but that's, of course, a matter of taste

General comments on the code base:
- [x] not sure if I am missing something, but I couldn't find any unittests
- [x] there are no community guidelines, i.e., how to contribute, report issues, or get help
- [x] the repository is huge (more than 800MB); there seem to be many left-over files that were likely committed accidentally and inflate the repository a lot:

```
057f58044e3c  8,9MiB modulo_python_package/modulo_vki/docs/build/doctrees/environment.pickle
36c717ad98e9  9,7MiB docs/build/doctrees/environment.pickle
59ddefba7ae2   10MiB docs/build/doctrees/environment.pickle
60a893a9b564   11MiB Matlab_Exercises/Exercise_1/Correlation_K.mat
db9b053e5e8f   11MiB Matlab_Exercises/Exercise_1/Psis_POD.mat
832e9ec09fa3   11MiB docs/build/doctrees/environment.pickle
b3807feb579d   12MiB docs/build/doctrees/environment.pickle
7e5d482fcd1a   12MiB docs/build/doctrees/environment.pickle
6aa60b1937ee   12MiB docs/build/doctrees/environment.pickle
04278c5c74d0   12MiB docs/build/.doctrees/environment.pickle
d086f7ad2fbb   12MiB docs/build/.doctrees/environment.pickle
56bf44113a2a   12MiB Matlab_Exercises/Exercise_1/Exercise_1.gif
a707d586d5a8   26MiB DATA_EXERCISES/Ex_4_TR_PIV_4.zip
962341982fa8   26MiB DATA_EXERCISES/Ex_4_TR_PIV_1.zip
f9d01aba6325   26MiB DATA_EXERCISES/Ex_4_TR_PIV_3.zip
50fc801c91fb   26MiB DATA_EXERCISES/Ex_4_TR_PIV_2.zip
dd5bbe4a65af   31MiB Matlab_Exercises/Exercise_3/CFD_Vortices_com.zip
166cd69026ef   31MiB Matlab_Exercises/Exercise_3/CFD_Vortices_com.zip
aca290d0bf05   34MiB Matlab_Exercises/Exercise_3/CFD_Vortices_com.zip
327fe3b48b5f   34MiB Matlab_Exercises/Exercise_3/CFD_Vortices_com.zip
b2cc8e5df694   37MiB Matlab_Exercises/Exercise_4/TR_PIV_2.zip
e088e4e6a4e6   37MiB Matlab_Exercises/Exercise_4/TR_PIV_1.zip
0105e5864394   38MiB Python_Exercises/Exercise_4/Exercise_4.gif
9173241ddc59   45MiB Matlab_Exercises/Exercise_4/TR_PIV_2.zip
46d85e5dffc2   45MiB Matlab_Exercises/Exercise_4/TR_PIV_1.zip
67eeac206484   67MiB DATA_EXERCISES/Ex_3_CFD_Vortices_com.zip
32e8d3ad6f8c   67MiB Matlab_Exercises/Exercise_3/CFD_Vortices_com.zip
a927a09bdbdf   82MiB Matlab_Exercises/Exercise_4/TR_PIV.zip
7d29b8a504f0   89MiB Exercise_4/TR_PIV.zip
```
To reproduce this output, run:
```
git rev-list --objects --all |   git cat-file --batch-check='%(objecttype) %(objectname) %(objectsize) %(rest)' |   sed -n 's/^blob //p' |   sort --numeric-sort --key=2 |   cut -c 1-12,41- |   $(command -v gnumfmt || echo numfmt) --field=2 --to=iec-i --suffix=B --padding=7 --round=nearest
```

Thanks for publishing your implementation. Please let me know if any of my comments are unclear.

Andre
