### 2024-07-02 mendezVKI/MODULO#28 @mengaldo
JOSS review
@mendezVKI, @philipcardiff, @AndreWeiner

- [x] The motivation and statement of need can be better stated (not only for fluid mechanics).
- [x] Some typos here and there that can be addressed.
- [x] I cannot find any tests; these should be part of the package.
- [x] It would be great to have instructions on how to install and run the package also in the GitHub repository, and not only on the external doc website.
- [x] Competing works are missing:
    1. Rogowski, Marcin, Brandon CY Yeung, Oliver T. Schmidt, Romit Maulik, Lisandro Dalcin, Matteo Parsani, and Gianmarco Mengaldo. "Unlocking massively parallel spectral proper orthogonal decompositions in the PySPOD package." Computer Physics Communications 302 (2024): 109246.
    2. Lario, A., Maulik, R., Schmidt, O.T., Rozza, G. and Mengaldo, G., 2022. Neural-network learning of SPOD latent dynamics. Journal of Computational Physics, 468, p.111475.

- [x] Is your library parallel? If not, that should be stated, stating how these may prevent running large-scale simulation of big data.
- [x] The size of the library seems not within GitHub guidelines. Some unintended files added?
