### 2024-07-02 #6753 @philipcardiff
> Hi @philipcardiff, I have now completed the checklist, thank you.
> If comments are required about the unchecked boxes, let me know.
> I opened an issue with my review.

Thank you, @mengaldo.

@mendezVKI: please feel free to ask for clarification on any outstanding points.

Review breadcrumb: https://github.com/mendezVKI/MODULO/issues/28
