+++ README.rst
@@ -1,0 +1,270 @@
+
+
+MODULO: MODal mULtiscale pOd
+-----------------------------
+
+**MODULO** is a modal decomposition package developed at the von Karman Institute for Fluid Dynamics (VKI). 
+It offers a wide range of decomposition techniques, enabling users to select the most suitable method based 
+on the specific physics of their problem and their desired outcomes. MODULO can natively handle large
+datasets efficiently, featuring a memory-saving option that partitions the data and computes the decomposition in 
+chunks (cite_ninni). Additionally, it supports non-uniform meshes through its weighted inner product formulation.
+
+While the discontinued MATLAB version of MODULO (cite_ninni) is accessible in the “Old_Matlab_Implementation” branch, 
+it is no longer maintained. The latest decomposition techniques are exclusively available in the current Python version.
+
+As a part of the MODULO project, we provide a series of lectures on data-driven modal decomposition, and its applications.
+These are available at the MODULO YouTube channel (https://www.youtube.com/@modulompod5682).
+
+
+.. contents:: Table of contents
+
+Modal decompositions
+--------------------
+Modal decompositions aim to describe the data as a linear combination of *modes*, obtained by projecting the data 
+onto a suitable set of basis. For instance, consider a matrix $D(x, t)$, where $x$ and $t$ are the spatial and temporal
+coordinates, respectively, the modal decomposition can be written as:
+
+.. math::
+    D(x_i, t_k) = \phi(x_i) \Sigma \psi(t_k)^T,
+
+where $\phi(x_i)$ is the spatial basis, $\psi(t_k)$ is the temporal basis, and $\Sigma$ is the modal coefficients. 
+Different decompositions employ different basis, such as prescribed Fourier basis ($$\psi_\mathcal{F}$$) for 
+the Discrete Fourier Transform (DFT), or data-driven basis, i.e. tailored on the dataset at hand, 
+for the Proper Orthogonal Decomposition (POD). 
+
+We refer to (mendez2022statistical, mendez2022generalizedmultiscalemodalanalysis, Mendez_2023) for an introduction to the topic.
+
+MODULO currently features the following decompositions: 
+- Discrete Fourier Transform (DFT) (briggs1995dft)
+- Proper Orthogonal Decomposition (POD) (sirovich1987turbulence, berkooz1993proper)
+- Multi-Scale Proper Orthogonal Decomposition (mPOD) (mendez2019multi)
+- Dynamic Mode Decomposition (DMD) (schmid2010dynamic)
+- Spectral Proper Orthogonal Decomposition (SPOD) (csieber2016spectral, towne2018spectral), note that the two are 
+    different formulations, and both are available in MODULO.
+- Kernel Proper Orthogonal Decomposition (KPOD) (mika1998kernel)
+
+We remind the curious reader to the respective references for a detailed description of each decomposition, and to the
+documentation for a practical guide on how to use them in MODULO.
+
+
+Release Notes
+-------------
+The latest version of MODULO (v2.0) includes the following updates:
+
+1. **Faster EIG/SVD algorithms**, using powerful randomized svd solvers from scikit_learn 
+    (see `here<https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.TruncatedSVD.html>`_ 
+    and `here <https://scikit-learn.org/stable/modules/generated/sklearn.utils.extmath.randomized_svd.html>`_.) 
+    It is now possible to select various options as "eig_solver" and "svd_solver", 
+    offering different trade-offs in terms of accuracy vs computational time.
+
[…]
