+++ README.rst
@@ -27,8 +27,8 @@ coordinates, respectively, the modal decomposition can be written as:
 .. math::
     D(x_i, t_k) = \phi(x_i) \Sigma \psi(t_k)^T,
 
-where $\phi(x_i)$ is the spatial basis, :math:`\psi(t_k)` is the temporal basis, and :math:`\Sigma` is the modal coefficients. 
-Different decompositions employ different basis, such as prescribed Fourier basis (:math:`\psi_\mathcal{F}`) for 
+where $\\phi(x_i)$ is the spatial basis, $\\psi(t_k)$ is the temporal basis, and $\\Sigma$ is the modal coefficients. 
+Different decompositions employ different basis, such as prescribed Fourier basis ($\\psi_\\mathcal{F}$) for 
 the Discrete Fourier Transform (DFT), or data-driven basis, i.e. tailored on the dataset at hand, 
 for the Proper Orthogonal Decomposition (POD). 
 
@@ -60,28 +60,28 @@ The latest version of MODULO (v2.0) includes the following updates:
 2. **Computation the POD directly via SVD**, using any of the four "svd_solver" options.
 This is generally faster but requires more memory.
 
-3. **Faster subscale estimators for the mPOD:** the previous version used the rank of the correlation matrix in each scale to define 
+1. **Faster subscale estimators for the mPOD:** the previous version used the rank of the correlation matrix in each scale to define 
 the number of modes to be computed in each portion of the splitting vector before assembling the full basis. 
 This is computationally very demanding. This estimation has been replaced by a 
 frequency-based threshold (i.e. based on the frequency bins within each portion) since one can show that the 
 frequency-based estimator is always more "conservative" than the rank-based estimator.
 
-4. **Major improvement on the memory saving option**: the previous version of modulo always required in input the matrix D. 
+1. **Major improvement on the memory saving option**: the previous version of modulo always required in input the matrix D. 
 Then, if the memory saving option was active, the matrix was partitioned and stored locally to free the RAM before computing the 
 correlation matrix (see [this tutorial by D. Ninni](https://www.youtube.com/watch?v=LclxO1WTuao)). 
 In the new version, it is possible to initialize a modulo object *without* the matrix D (see exercise 5 in the examples). 
 Instead, one can create the partitions without loading the matrix D.
 
-5. **Implementation of Dynamic Mode Decomposition (DMD)** from (Schmid, P.J 2010)
+1. **Implementation of Dynamic Mode Decomposition (DMD)** from (Schmid, P.J 2010)
 
-6. **Implementation of the two Spectral POD formulations**, namely the one from (Sieber et al 2016), 
+2. **Implementation of the two Spectral POD formulations**, namely the one from (Sieber et al 2016), 
    and the one from (Towne et al 2018).
 
-7. **Implementation of a kernel version of the POD**, in which the correlation matrix is replaced by a kernel matrix. 
+3. **Implementation of a kernel version of the POD**, in which the correlation matrix is replaced by a kernel matrix. 
 This is described in Lecture 15 of the course `Hands on Machine Learning for Fluid dynamics 2023 <https://www.vki.ac.be/index.php/events-ls/events/eventdetail/552/-/online-on-site-hands-on-machine-learning-for-fluid-dynamics-2023>`_. 
 We refer also to: `https://arxiv.org/abs/2208.07746`. 
 
-8. **Implementation of a formulation for non-uniform meshes**, using a weighted matrix for all the relevant inner products. 
+1. **Implementation of a formulation for non-uniform meshes**, using a weighted matrix for all the relevant inner products. 
 This is currently available only for POD and mPOD but allows for handling data produced from CFD simulation without resampling on a uniform grid (see exercise 4). 
 It can be used both with and without the memory-saving option.
 
