+++ README.rst
@@ -7,6 +7,8 @@ MODULO: MODal mULtiscale pOd
     :target: https://modulo.readthedocs.io/en/latest/?badge=latest
     :alt: Documentation Status
 
+.. _image-link: https://modulo.readthedocs.io/en/latest/_images/modulo_logo.png
+
 **MODULO** is a modal decomposition package developed at the von Karman Institute for Fluid Dynamics (VKI). 
 It offers a wide range of decomposition techniques, enabling users to select the most suitable method based 
 on the specific physics of their problem and their desired outcomes. MODULO can natively handle large
