+++ README.rst
@@ -7,12 +7,11 @@ MODULO: MODal mULtiscale pOd
     :target: https://modulo.readthedocs.io/en/latest/?badge=latest
     :alt: Documentation Status
 
-.. |modulo_logo| image:: https://modulo.readthedocs.io/en/latest/_images/modulo_logo.png
+.. image:: https://modulo.readthedocs.io/en/latest/_images/modulo_logo.png
    :alt: Modulo Logo
    :width: 400
    :align: center
 
-|modulo_logo|
 
 **MODULO** is a modal decomposition package developed at the von Karman Institute for Fluid Dynamics (VKI). 
 It offers a wide range of decomposition techniques, enabling users to select the most suitable method based 
