+++ paper/paper.bib
@@ -109,3 +109,22 @@ author = {Gianmarco Mengaldo and Romit Maulik},
   year = {2023},
   note = {Accessed on April 27, 2023}
 }
+
+@article{rogowski2024unlocking,
+  title={Unlocking massively parallel spectral proper orthogonal decompositions in the PySPOD package},
+  author={Rogowski, Marcin and Yeung, Brandon CY and Schmidt, Oliver T and Maulik, Romit and Dalcin, Lisandro and Parsani, Matteo and Mengaldo, Gianmarco},
+  journal={Computer Physics Communications},
+  volume={302},
+  pages={109246},
+  year={2024},
+  publisher={Elsevier}
+}
+@article{demo2018pydmd,
+  title={PyDMD: Python dynamic mode decomposition},
+  author={Demo, Nicola and Tezzele, Marco and Rozza, Gianluigi},
+  journal={Journal of Open Source Software},
+  volume={3},
+  number={22},
+  pages={530},
+  year={2018}
+}
