### 2024-07-10 1bb4ba3
sorted tutorials and plots
