### 2024-07-10 e7bdb8a
sorted tutorials and plots
