+++ modulo_vki/test/run_all_tests.py
@@ -1,0 +1,8 @@
+import os
+
+os.system('python test_modulo.py')
+os.system('python utils/test_read_db.py')
+os.system('python utils/test_utils.py')
+os.system('python core/test_k_matrix.py')
+os.system('python core/test_pod_space.py')
+
