### 2024-07-17 30e7eb5
Minor bug fixes to memory saving in _pod_space and edits in tests
