### 2024-07-24 3b00724
Update README.rst
minor adjustments readme
