+++ paper/paper.md
@@ -51,7 +51,7 @@ Dimensionality reduction is an essential tool in processing large datasets, enab
 This work proposes a major upgrade of the software package MODULO (MODal mULtiscale pOd,[@ninni_modulo_2020]), which was designed to perform Multiscale Proper Orthogonal Decomposition (mPOD)[@mendez_balabane_buchlin_2019]. In addition to implementing the classic Fourier Transform (DFT) and Proper Orthogonal Decomposition (POD), MODULO now also allows for computing Dynamic Mode Decomposition (DMD) [@schmid_2010] as well as the Spectral POD by [@sieber_paschereit_oberleithner_2016], the Spectral POD by [@Towne_2018] and a generalized kernel-based decomposition akin to kernel PCA [@mendez_2023]. All algorithms are wrapped in a ‘SciKit’-like Python API which allows computing all decompositions in one line of code. Documentation, exercises, and video tutorials are also provided to offer a primer on data drive modal analysis.
 
 # Statement of Need
-As extensively illustrated in recent reviews [@mendez_2023], [@Taira2020], all modal decompositions can be seen as special kinds of matrix factorizations. The matrix being factorized collects (many) snapshots (samples) of a high-dimensional variable. The factorization provides a basis for the column and the row spaces of the matrix, to identify the most essential patterns (modes) according to a certain criterion. In what follows, we will refer to common terminologies in fluid dynamics. Still, the reader should be aware that these tools are general and can be applied to any high-dimensional dataset, to identify its most important features and to build reduced-order models. In the common arrangement encountered in fluid dynamics, the basis for the column space is a set of ‘spatial structures’ while the basis for the row space is a set of `temporal structures'. These are paired by a scalar which defines their relative importance. The POD, closely related to Principal Component Analysis, yields modes with the highest energy (variance) content and, in addition, guarantees their orthonormality by construction.
+As extensively illustrated in recent reviews [@mendez_2023], [@Taira2020], all modal decompositions can be seen as special kinds of matrix factorizations. The matrix being factorized collects (many) snapshots (samples) of a high-dimensional variable. The factorization provides a basis for the column and the row spaces of the matrix, to identify the most essential patterns (modes) according to a certain criterion. In what follows, we will refer to common terminologies in fluid dynamics. Nevertheless, it is worth stressing that these tools can be applied to any high-dimensional dataset to identify patterns and build reduced-order models. In the common arrangement encountered in fluid dynamics, the basis for the column space is a set of ‘spatial structures’ while the basis for the row space is a set of `temporal structures'. These are paired by a scalar which defines their relative importance. The POD, closely related to Principal Component Analysis, yields modes with the highest energy (variance) content and, in addition, guarantees their orthonormality by construction.
 In the DFT, as implemented in MODULO, modes are defined to evolve as orthonormal complex exponential in time. This implies that the associated frequencies are integer multiples of a fundamental tone. The DMD generalizes the DFT by releasing the constraint of orthogonality and considering complex frequencies, i.e., modes that can potentially vanish or decay. 
[…]
