+++ paper/paper.bib
@@ -128,3 +128,10 @@ author = {Gianmarco Mengaldo and Romit Maulik},
   pages={530},
   year={2018}
 }
+@article{Mendez_Balabane_Buchlin_2019, 
+title={Multi-scale proper orthogonal decomposition of complex fluid flows}, 
+volume={870}, DOI={10.1017/jfm.2019.212}, 
+journal={Journal of Fluid Mechanics}, 
+author={Mendez, M. A. and Balabane, M. and Buchlin, J.-M.}, 
+year={2019}, 
+pages={988–1036}} 
