### 2024-07-24 mendezVKI/MODULO#28 @mendezVKI
Dear @mengaldo , @philipcardiff, @AndreWeiner

Thank you very much for your thorough review. We believe we have addressed most of the points, though we encountered some difficulties with the one concerning the size of the repository.

Here is the summary of the updates:

We added Unit tests.

The paper now includes citations to similar packages and one sentence stating the potential applications outside fluid mechanics.

The readme has been updated with instructions for installation and usage, as well as more context and references.

Concerning the parallel implementation: currently MODULO does not offer additional parallel computing capabilities beyond those naturally provided by NumPy. This is now explicitly stated in the Readme. Further parallel implementations are planned for future developments. Note, however, that the memory saving feature does allow to process quite large datasets effectively.

A partially open point is the one concerning the size of the repository.

The repository's large size is mainly due to old files committed some time ago. This issue is specific to Git, meaning that users accessing the package from the PyPI repository will not be affected.  More specifically, these files are inside the .git folder and the only way to delete these, as far as we know, is to re-write the history of the repository. This means that we would have to delete the repository and create a new one with a cleaner history.

However, we are afraid that the people who forked the repository (we have currently 24) will not be able to push back their developments.  We would appreciate any suggestion on how to clean this folder in another way.

Nevertheless, to circumvent the problem, at this stage we encourage potential developers to shallow-clone the repository with the command:

git clone --depth 1 <repository_url>

This is written both in the contribution guideline and in the readme and allows to download all the relevant codes with less than 15MB. We highlight that shallow cloning the repository does not limit users in the collaboration (e.g. pull requests, issues, merges), as the commits involving large files are referring to the preliminary version of the software and are no longer relevant.
