### 2024-07-24 mendezVKI/MODULO#27 @mendezVKI
Dear @AndreWeiner

Thank you very much for your commitment to the review of our work. We addressed all points, except 2 of them.

Solved issues:

We run several grammar checks and it seems good to us.

We added a citation to other projects implementing DMD, SPOD, etc.

We correct all issues in the tutorials

We added Unit tests

We added community guidelines.

Open points :

Including an example with mPOD and comparison in the paper. We have various tutorials comparing mPOD with other decompositions, highlighting the main advantages etc. This is also largely discussed in various papers ([here](https://www.cambridge.org/core/journals/journal-of-fluid-mechanics/article/abs/multiscale-proper-orthogonal-decomposition-of-complex-fluid-flows/D078BD2873B1C30B6DD9016E30B62DA8) and [here](https://arxiv.org/abs/2001.01971)) and chapters ([here](https://www.cambridge.org/core/books/abs/datadriven-fluid-mechanics/generalized-and-multiscale-modal-analysis/F6C342E80A83D68756CEC1492848C6E1)). We do not see the point of adding these also in the paper.

Size of the repository: please see the reply to the comment in [#issue28](https://github.com/mendezVKI/MODULO/issues/28). Indeed, we have a small problem there and are unsure about how to proceed. Suggestions are welcome. Nevertheless, we have found a way to circumvent the issue and described it in the readme.
