### 2024-07-24 mendezVKI/MODULO#28 @mengaldo
"Concerning the parallel implementation: currently MODULO does not offer additional parallel computing capabilities beyond those naturally provided by NumPy. This is now explicitly stated in the Readme. Further parallel implementations are planned for future developments. Note, however, that the memory saving feature does allow to process quite large datasets effectively."

@mendezVKI: Quite large datasets is relative to the application.
If you do not have distributed capabilities, it is likely that you cannot run say a 100TB in a single node.
Am I misunderstanding something here?
