### 2024-07-24 mendezVKI/MODULO#28 @mendezVKI
Sure, you're right: size is very much application dependent and 100 TB sounds off-scale for us.
But the main issue in decomposing such a large case is more memory requirements than parallel capabilities.

With the the memory saving, you could in principle process a dataset of that 100 TB in MODULO, provided that you partition it into chunks sufficiently small to fit in your RAM. On a good desktop with 128 GB of RAM, you would need about 800 partitions. That would of course take quite a lot of time,, most of which being spend in I/O operation to save intermediate results.

The potential gains from parallel computing are still uncertain for us. The most computationally intensive operations are inner products and eigenvalue computations, which NumPy handles in a quasi-parallel manner. On Windows, these operations typically utilize around 70% all the available processors, whereas, on Linux, eigenvalue computations often use 100% of processors by default, with other operations (such as inner products and I/O) utilizing around 60%. However, for the scale you're discussing, Python might not be the most suitable tool.

To give you an order of magnitude for our most common applications in image velocimetry, in planar cases databases of the order of GB are very large. In 3D velocimetry, we might reach something of the order of dozens GB. These can all run quite easily. The largest we have worked are on the order of 50-100 GB and could be easily digested on my 32GB RAM laptop thanks to the memory saving.
