### 2024-07-24 mendezVKI/MODULO#28 @mengaldo
Are you suggesting that this is feasible?
_"With the the memory saving, you could in principle process a dataset of that 100 TB in MODULO, provided that you partition it into chunks sufficiently small to fit in your RAM. On a good desktop with 128 GB of RAM, you would need about 800 partitions. That would of course take quite a lot of time,, most of which being spend in I/O operation to save intermediate results."_

I would recommend to take the geophysical example from this paper (~100TB), and demonstrate that this is the case: https://arxiv.org/abs/2309.11808

Also, "_which NumPy handles in a quasi-parallel manner_". As far as I know, this is shared-memory parallelism. To go to distributed parallelism you would need to use an mpi package.
