### 2024-07-24 mendezVKI/MODULO#28 @mendezVKI
Indeed, there is quite a lot of work to do for improving the parallelism.

Anyway, yes: 100 TB could be in principle possible depending on the proportion of the number of snapshot versus the weight of the snapshot.

You can see all the details to compute how much memory you need with the memory saving in [this video](https://www.youtube.com/watch?v=LclxO1WTuao).

Your data has relatively small snapshots but too many of them for the current implementation in MODULO.
I understand you have  n_s=38 414 880 and n_t=727584.

With the memory saving you do not need to store all the n_s x n_t points, but you need to store the temporal correlation matrix K  and the temporal basis PSI (this is much smaller). As long as n_t^2 is much smaller than n_s x n_t the memory saving helps a lot and can enable digesting huge datasets. Unfortunately this is not your case.

In your case, K is of size 727584 x 727584, which would require about 1.92 TB of memory. Unless you have a computer with that much RAM we cannot do it with the current MODULO.

Roughly speaking, if you have 128 GB of RAM you could deal with datasets having about n_t = 185,363 snapshots.
To consider datasets of the order of 100 TB means that you have snapshots of about n_s=148,289,744.69, that is about 0,55 GB per snapshot. You could have chunks of about n_t’=230 to stay below 128GB per chunk and proceed chunk by chunk. It would take a lot of time, but it is not impossible.

So the point is not 100TB or less. Its about how much is n_s and n_t.

While I appreciate you sharing your work with us, dealing with such datasets is beyond the scope of our current release. We don't have the time or resources to pursue those developments at the moment, and they are not aligned with our current focus, especially since we don't work with data at that scale. We've provided all the necessary information, tutorials, and videos to help assess whether the current version of MODULO is suitable for a given problem. I'm confident that it remains valuable for addressing issues relevant to a large community.
