### 2024-07-24 mendezVKI/MODULO#28 @mengaldo
Thank you for the detailed response.

My point is that it seems that you suggest that you can run easily very large datasets.
This is usually not achievable or scalable without a distributed parallel implementation.

For instance, how long would it take using chunking, if it was feasible?

I believe you have a few intermediate steps on the I/O side that would make the runtime of large datasets like the one above very large?

In fact, I/O becomes one of the critical bottlenecks that is not easy to handle if you go to datasets on the order of several terabytes or more.

I believe the library is of interest to the community, but the above limitations may be better highlighted.
