### 2024-07-25 4940906
add cost estimates
