### 2024-07-25 e72677e
add cost estimates
