### 2024-07-25 6a3d205
add cost estimates
