+++ README.rst
@@ -209,30 +209,30 @@ Computational Cost Estimates
 To guide the user in choosing the appropriate setting for the specific dataset at hand, we provide a rough estimate of the computational 
 cost of the MODULO decompositions, with and without the memory-saving option.
 
-Given a dataset of size :math:`n_s \times n_t`, we consider the computation of :math:`n_r` modes. When using the memory saving option, we refer to 
-:math:`n_t' = n_t / n_p` as the number of time steps in each partition, and to :math:`n_s' = n_s / n_p` as the number of spatial points in each partition.
+Given a dataset of size ``n_s \times n_t``, we consider the computation of ``n_r`` modes. When using the memory saving option, we refer to 
+``n_t' = n_t / n_p`` as the number of time steps in each partition, and to ``n_s' = n_s / n_p`` as the number of spatial points in each partition.
 
 .. list-table::
     :widths: 25 25 25 25
     :header-rows: 1
 
     * - 
-      - Phase 1: :math:`D`
-      - Phase 2: :math:`K`
-      - Phase 3: :math:`\Psi`
-      - Phase 4: :math:`\Phi`
+      - Phase 1: ``D``
+      - Phase 2: ``K``
+      - Phase 3: ``\Psi``
+      - Phase 4: ``\Phi``
     * - No Memory Saving
-      - :math:`n_s \times n_t`
-      - :math:`n_t^2`
-      - :math:`n_t^2 + n_t \times n_r`
-      - :math:`n_s \times n_t + n_t \times n_r + n_s \times n_r`
+      - ``n_s \times n_t``
+      - ``n_t^2``
+      - ``n_t^2 + n_t \times n_r``
+      - ``n_s \times n_t + n_t \times n_r + n_s \times n_r``
     * - Memory Saving
       - /
-      - :math:`n_s \times n_t' + n_t' \times n_t'`
-      - :math:`n_t^2 + r \cdot n_t`
-      - :math:`(n_s \times n_t') + n_s' \times n_t + n_s \times n_r`
+      - ``n_s \times n_t' + n_t' \times n_t'``
+      - ``n_t^2 + r \cdot n_t``
+      - ``(n_s \times n_t') + n_s' \times n_t + n_s \times n_r``
 
-As shown in the table, the computational cost of the decomposition is dominated by the computation of the correlation matrix :math:`K` in Phase 2. The 
+As shown in the table, the computational cost of the decomposition is dominated by the computation of the correlation matrix ``K`` in Phase 2. The 
 memory-saving option significantly reduces the computational cost of this phase, as the correlation matrix is computed for each partition, rather than
 for the entire dataset.
 
