+++ README.rst
@@ -227,7 +227,7 @@ Given a dataset of size ``n_s \times n_t``, we consider the computation of ``n_r
       - ``n_t^2 + n_t \times n_r``
       - ``n_s \times n_t + n_t \times n_r + n_s \times n_r``
     * - Memory Saving
-      - /
+      - ``/``
       - ``n_s \times n_t' + n_t' \times n_t'``
       - ``n_t^2 + r \cdot n_t``
       - ``(n_s \times n_t') + n_s' \times n_t + n_s \times n_r``
