### 2024-07-25 c0c3332
add cost estimates
