### 2024-07-25 f0190d1
add cost estimates
