### 2024-07-25 90a4e1c
add cost estimates
