### 2024-07-25 0459562
add cost estimates
