### 2024-07-25 902f1e6
add cost estimates
