### 2024-07-25 5696148
add cost estimates
