### 2024-07-25 c8b1ac4
add cost estimates
