### 2024-07-25 17aa8ee
add cost estimates
