### 2024-07-25 d2f0e7f
add cost estimates
