### 2024-07-25 3527e9e
add cost estimates
