### 2024-07-25 410e296
add cost estimates
