### 2024-07-25 1847e55
add cost estimates
