### 2024-07-25 8cce371
add cost estimates
