### 2024-07-26 ae80c7d
Update README.rst
readme update with table of memory requirements
