### 2024-07-26 mendezVKI/MODULO#28 @mendezVKI
We removed the sentence ``quite large datasets effectively'' and added a section with a table to compute the memory requirement for a given dataset using the memory saving feature or not.

These are only based on memory requirements. How much time it would take to deal with a given dataset while facing the I/O bottleneck depends on what hardware is being used and we do not know how to estimate it in general. But anyway, here we are talking about "it is possible" vs "it is not possible", not about "how fast" it would be.
