### 2024-07-29 #6753 @AndreWeiner
@mendezVKI In line 37 of the updated article, you write *modes that can potentially vanish or decay*. I assume you meant to write something like *can decay or grow* or *can vanish or explode*.

@philipcardiff Besides the minor remark above, I consider the review complete from my side. My remarks have been addressed by the authors.

Best, Andre
