### 2024-07-29 #6753 @philipcardiff
> @mendezVKI In line 37 of the updated article, you write _modes that can potentially vanish or decay_. I assume you meant to write something like _can decay or grow_ or _can vanish or explode_.
>
> @philipcardiff Besides the minor remark above, I consider the review complete from my side. My remarks have been addressed by the authors.
>
> Best, Andre

Thanks, Andre!
