+++ paper/paper.md
@@ -52,7 +52,7 @@ This work proposes a major upgrade of the software package MODULO (MODal mULtisc
 
 # Statement of Need
 As extensively illustrated in recent reviews [@mendez_2023], [@Taira2020], all modal decompositions can be seen as special kinds of matrix factorizations. The matrix being factorized collects (many) snapshots (samples) of a high-dimensional variable. The factorization provides a basis for the column and the row spaces of the matrix, to identify the most essential patterns (modes) according to a certain criterion. In what follows, we will refer to common terminologies in fluid dynamics. Nevertheless, it is worth stressing that these tools can be applied to any high-dimensional dataset to identify patterns and build reduced-order models [@Mendez_Balabane_Buchlin_2019]. In the common arrangement encountered in fluid dynamics, the basis for the column space is a set of ‘spatial structures’ while the basis for the row space is a set of `temporal structures'. These are paired by a scalar which defines their relative importance. The POD, closely related to Principal Component Analysis, yields modes with the highest energy (variance) content and, in addition, guarantees their orthonormality by construction.
-In the DFT, as implemented in MODULO, modes are defined to evolve as orthonormal complex exponential in time. This implies that the associated frequencies are integer multiples of a fundamental tone. The DMD generalizes the DFT by releasing the constraint of orthogonality and considering complex frequencies, i.e., modes that can potentially vanish or decay. 
+In the DFT, as implemented in MODULO, modes are defined to evolve as orthonormal complex exponential in time. This implies that the associated frequencies are integer multiples of a fundamental tone. The DMD generalizes the DFT by releasing the constraint of orthogonality and considering complex frequencies, i.e., modes that can potentially vanish or explode. 
 Both the constraint of energy optimality and harmonic modes can lead to poor performances in terms of convergence and feature detection. This motivated the development of hybrid methods such as the Spectral POD by [@Towne_2018], Spectral POD by [@sieber_paschereit_oberleithner_2016], and Multiscale Proper Orthogonal Decomposition (mPOD)[@mendez_balabane_buchlin_2019]. The first can be seen as an optimally averaged DMD while the second consists in bridging POD and DFT with the use of a filtering operation. Both SPODs assume statistically stationary data and are designed to identify harmonic (or quasi-harmonic) modes. The mPOD combines POD with Multi-resolution Analysis (MRA), to provide modes that are optimal within a prescribed frequency band. The mPOD modes are thus spectrally less narrow than those obtained by the SPODs, but this allows for localizing them in time (i.e. potentially having compact support in time). 
 Finally, recent developments in nonlinear methods such as kernel PCA and their applications to fluid dynamics (see [@mendez_2023]) have motivated the interest in the connection between nonlinear methods and the most general Karhunen–Loeve expansion (KL). This generalizes the POD as the decomposition of data onto the eigenfunction of a kernel function (the POD being a KL for the case of linear kernel). 
 
