### 2024-08-14 #6753 @philipcardiff
Hi @mendezVKI, yes, correct; that is my understanding.

@AndreWeiner recommends the submission for publication, while, from @mengaldo's comment above, he is happy to clarify any queries you have about his unchecked points.
