### 2024-10-08 #6753 @mendezVKI
Dear @philipcardiff ,

All issues from reviewers were closed as of August 20th.

@mengaldo closed his issues but left some boxes unchecked and since then we have asked multiple times (last time on 1 month ago) to have some more info but we have no response.

It's now been quite a while since we finished the revision.

Is there anything else we can do to move forward?
