### 2024-10-08 #6753 @philipcardiff
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.1002/nme.255 is OK
> - 10.1017/jfm.2016.103 is OK
> - 10.1007/s11340-014-9858-2 is OK
> - 10.1016/j.softx.2020.100622 is OK
> - 10.1017/jfm.2019.212 is OK
> - 10.1017/S0022112010001217 is OK
> - 10.1017/jfm.2018.283 is OK
> - 10.1088/1361-6501/acaffe is OK
> - 10.2514/1.J058462 is OK
> - 10.21105/joss.02862 is OK
> - 10.1017/jfm.2019.212 is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: Proper Orthogonal Decomposition
> - No DOI given, and none found for title: Spectral Proper Orthogonal Decomposition
>
> ❌ MISSING DOIs
>
> - 10.1016/j.cpc.2024.109246 may be a valid DOI for title: Unlocking massively parallel spectral proper ortho...
> - 10.21105/joss.00530 may be a valid DOI for title: PyDMD: Python dynamic mode decomposition
>
> ❌ INVALID DOIs
>
> - https://doi.org/10.21105/joss.00530 is INVALID because of 'https://doi.org/' prefix
> ```

Hi @mendezVKI , please check and fix the missing and invalid DOIs. Thanks!

You will also see that I proposed some minor English edits to the paper above via a pull request.
