+++ paper/paper.bib
@@ -74,15 +74,16 @@ eprint = { https://doi.org/10.2514/1.J058462}
 }
 
 @article{py_DMD,
-  Author = {Demo, Nicola and Tezzele, Marco and Rozza, Gianluigi},
-  Title = {{PyDMD: Python Dynamic Mode Decomposition}},
-  Journal = {The Journal of Open Source Software},
-  Volume = {3},
-  Number = {22},
-  Pages = {530},
-  Year = {2018},
-  Doi = {https://doi.org/10.21105/joss.00530}
+  title={PyDMD: Python dynamic mode decomposition},
+  author={Demo, Nicola and Tezzele, Marco and Rozza, Gianluigi},
+  journal={Journal of Open Source Software},
+  doi = {10.21105/joss.00530},
+  volume={3},
+  number={22},
+  pages={530},
+  year={2018}
 }
+
 
 @misc{py_POD,
   title = {Proper Orthogonal Decomposition},
@@ -111,14 +112,22 @@ author = {Gianmarco Mengaldo and Romit Maulik},
 }
 
 @article{rogowski2024unlocking,
-  title={Unlocking massively parallel spectral proper orthogonal decompositions in the PySPOD package},
-  author={Rogowski, Marcin and Yeung, Brandon CY and Schmidt, Oliver T and Maulik, Romit and Dalcin, Lisandro and Parsani, Matteo and Mengaldo, Gianmarco},
-  journal={Computer Physics Communications},
-  volume={302},
-  pages={109246},
-  year={2024},
-  publisher={Elsevier}
+title = {Unlocking massively parallel spectral proper orthogonal decompositions in the PySPOD package},
+journal = {Computer Physics Communications},
+volume = {302},
+pages = {109246},
+year = {2024},
+issn = {0010-4655},
+doi = {10.1016/j.cpc.2024.109246},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465524001693},
+author = {Marcin Rogowski and Brandon C.Y. Yeung and Oliver T. Schmidt and Romit Maulik and Lisandro Dalcin and Matteo Parsani and Gianmarco Mengaldo},
+keywords = {Spectral proper orthogonal decomposition, SPOD, Parallel, Distributed, MPI, Modal decomposition, Dynamical systems},
+abstract = {We propose a parallel (distributed) version of the spectral proper orthogonal decomposition (SPOD) technique. The parallel SPOD algorithm distributes the spatial dimension of the dataset preserving time. This approach is adopted to preserve the non-distributed fast Fourier transform of the data in time, thereby avoiding the associated bottlenecks. The parallel SPOD algorithm is implemented in the PySPOD library and makes use of the standard message passing interface (MPI) library, implemented in Python via mpi4py. An extensive performance evaluation of the parallel package is provided, including strong and weak scalability analyses. The open-source library allows the analysis of large datasets of interest across the scientific community. Here, we present applications in fluid dynamics and geophysics, that are extremely difficult (if not impossible) to achieve without a parallel algorithm. This work opens the path toward modal analyses of big quasi-stationary data, helping to uncover new unexplored spatio-temporal patterns.
+Program summary
[…]
