### 2024-10-14 #6753 @lorenzoschena
Hi @philipcardiff,

We have merged your PR and modified the invalid DOIs. Does everything look good now?

Best,
