### 2024-10-16 #6753 @lorenzoschena
Hi @philipcardiff ,

- [x] check authors and affiliations (paper.md, CITATION.cff)
- [x] release with latest changes: 2.0.7
- [x]  Zenodo DOI: 10.5281/zenodo.13939520
- [x] Authors list verification OK
- [x] MIT License is consistently applied in all software releases.

Let me know if something else's needed from our side.

Cheers,
