### 2024-10-16 #6753 @lorenzoschena
Hi @philipcardiff, it should show up now. Current releases substitutes v2.0 with some very minor fixes (mostly related to PyPI publishing)
