### 2024-10-17 1763fa8
Updates to JOSS paper
Fixes some reference commands
