### 2024-10-17 e854e34
Fixes JOSS paper bibilography
removes duplicate entry and adds missing data
