+++ paper/paper.bib
@@ -13,13 +13,22 @@
   publisher = {Wiley},
 }
 
-@article{sieber_paschereit_oberleithner_2016, title={Spectral proper orthogonal decomposition}, volume={792}, DOI={10.1017/jfm.2016.103}, journal={Journal of Fluid Mechanics}, publisher={Cambridge University Press}, author={Sieber, Moritz and Paschereit, C. Oliver and Oberleithner, Kilian}, year={2016}, pages={798–828}}
+@article{sieber_paschereit_oberleithner_2016, 
+  title={Spectral proper orthogonal decomposition}, 
+  volume={792}, 
+  DOI={10.1017/jfm.2016.103}, 
+  journal={Journal of Fluid Mechanics}, 
+  publisher={Cambridge University Press}, 
+  author={Sieber, Moritz and Paschereit, C. Oliver and Oberleithner, Kilian}, 
+  year={2016}, 
+  pages={798–828}
+}
 
 @article{berkooz_proper_1993,
 author = {Ranc, Nicolas and Blanche, Antoione and Ryckelynck, D. and Chrysochoos, Andre},
 year = {2014},
-month = {04},
-pages = {},
+month = apr,
+pages = {725--739},
 title = {POD Preprocessing of IR Thermal Data to Assess Heat Source Distributions},
 volume = {55},
 journal = {Experimental Mechanics},
@@ -27,7 +36,7 @@ doi = {10.1007/s11340-014-9858-2}
 }
 
 @article{ninni_modulo_2020,
-author = {Ninni, Davide and Mendez, Miguel},
+author = {Ninni, Davide and Mendez, Miguel A.},
 year = {2020},
 month = {12},
 pages = {100622},
@@ -37,7 +46,16 @@ journal = {SoftwareX},
 doi = {10.1016/j.softx.2020.100622}
 }
 
-@article{mendez_balabane_buchlin_2019, title={Multi-scale proper orthogonal decomposition of complex fluid flows}, volume={870}, DOI={10.1017/jfm.2019.212}, journal={Journal of Fluid Mechanics}, publisher={Cambridge University Press}, author={Mendez, M. A. and Balabane, M. and Buchlin, J.-M.}, year={2019}, pages={988–1036}}
+@article{mendez_balabane_buchlin_2019, 
+  title={Multi-scale proper orthogonal decomposition of complex fluid flows}, 
+  volume={870}, 
+  DOI={10.1017/jfm.2019.212}, 
+  journal={Journal of Fluid Mechanics}, 
+  publisher={Cambridge University Press}, 
+  author={Mendez, Miguel A. and Balabane, M. and Buchlin, J.-M.}, 
+  year={2019}, 
+  pages={988–1036}
+}
 
 @article{schmid_2010, 
 title={Dynamic mode decomposition of numerical and experimental data},
@@ -45,15 +63,27 @@ title={Dynamic mode decomposition of numerical and experimental data},
  DOI={10.1017/S0022112010001217}, 
  journal={Journal of Fluid Mechanics}, 
  publisher={Cambridge University Press},
- author={Schmid, Peter J.}, year={2010}, pages={5–28}}
+ author={Schmid, Peter J.}, 
+ year={2010}, 
+ pages={5–28}
+}
 
-@article{Towne_2018, title={Spectral proper orthogonal decomposition and its relationship to dynamic mode decomposition and resolvent analysis}, volume={847}, DOI={10.1017/jfm.2018.283}, journal={Journal of Fluid Mechanics}, publisher={Cambridge University Press}, author={Towne, Aaron and Schmidt, Oliver T. and Colonius, Tim}, year={2018}, pages={821–867}}
+@article{Towne_2018, 
[…]
