### 2024-10-17 97af41b
Fixes duplicate reference in JOSS paper
