### 2024-10-17 #6753 @kyleniemeyer
Hi @mendezVKI, can you review and merge this pull request? https://github.com/mendezVKI/MODULO/pull/32

I fixed some of the reference commands in the paper, and also found a duplicate entry.
